+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021-03-05
+
+### Added
+- Support for Google AdMob Native Ads.
+- Video player can now be set to fit a specific size or to full bleed.
+
+### Changed
+- iOS versions 11 through iOS 14.4 now supported.
+- The channel video sources now uses channel encoded ID instead of the channel name.
+
+### Removed
+- Publisher ID is no longer required or supported by the SDK.
+
 ## [0.2.0] - 2021-02-15
 
 ### Added